Totoro: added check in case setDB is None
git-svn-id: https://svn.sdss.org/repo/sdss/sdss_python_module/branches/josedev@64770 acfa968b-e9c4-4993-a9e6-846a59bbf143

+++ dbclasses/plate_utils.py
@@ -443,8 +443,9 @@ def applyArrangement(plate, arrangement):
                     exp.mangadbExposure.set_pk = None
                     if setPK is not None:
                         setDB = session.query(db.mangaDB.Set).get(setPK)
-                        session.delete(setDB)
-                        session.flush()
+                        if setDB is not None:
+                            session.delete(setDB)
+                            session.flush()
 
                 session.flush()
